requirements.txt
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
gTTS==2.5.4
pydub==0.25.1
Pillow==10.4.0
tqdm==4.66.4
